# app/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "PtLinker API",
        "description": "PT站点爬虫管理API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/tasks": {
            "post": {
                "tags": [
                    "tasks",
                    "tasks"
                ],
                "summary": "创建任务",
                "description": "创建任务\n\nArgs:\n    site_id: 指定站点ID（可选）\n    create_for_all_sites: 是否为所有站点创建任务\n    db: 数据库会话\n    site_manager: 站点管理器\n    queue_manager: 队列管理器\n    \nReturns:\n    List[TaskResponse]: 创建的任务列表\n    \nNote:\n    - 如果指定了 site_id，则只为该站点创建任务\n    - 如果设置了 create_for_all_sites=True，则为所有启用的站点创建任务\n    - 如果两者都未指定，则返回错误\n    - 单个站点创建失败不会影响其他站点的任务创建",
                "operationId": "create_task_api_v1_tasks_post",
                "parameters": [
                    {
                        "name": "site_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Site Id"
                        }
                    },
                    {
                        "name": "create_for_all_sites",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "description": "是否为所有站点创建任务",
                            "default": false,
                            "title": "Create For All Sites"
                        },
                        "description": "是否为所有站点创建任务"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TaskResponse"
                                    },
                                    "title": "Response Create Task Api V1 Tasks Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/tasks/{task_id}": {
            "get": {
                "tags": [
                    "tasks",
                    "tasks"
                ],
                "summary": "获取任务信息",
                "description": "获取指定任务的信息\n\nArgs:\n    task_id: 任务ID\n    db: 数据库会话\n    \nReturns:\n    TaskResponse: 任务信息\n    \nRaises:\n    HTTPException: 当任务不存在时",
                "operationId": "get_task_api_v1_tasks__task_id__get",
                "parameters": [
                    {
                        "name": "task_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Task Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "tasks",
                    "tasks"
                ],
                "summary": "取消任务",
                "description": "取消指定的任务\n\nArgs:\n    task_id: 任务ID\n    db: 数据库会话\n    queue_manager: 队列管理器\n    \nReturns:\n    dict: 操作结果\n    \nRaises:\n    HTTPException: 当任务不存在或无法取消时",
                "operationId": "cancel_task_api_v1_tasks__task_id__delete",
                "parameters": [
                    {
                        "name": "task_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Task Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response Cancel Task Api V1 Tasks  Task Id  Delete"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/tasks/": {
            "get": {
                "tags": [
                    "tasks",
                    "tasks"
                ],
                "summary": "获取任务列表",
                "description": "获取任务列表，支持按站点和状态筛选\n\nArgs:\n    site_id: 站点ID（可选）\n    status: 任务状态（可选）\n    limit: 返回数量限制\n    db: 数据库会话\n    \nReturns:\n    List[TaskResponse]: 任务列表",
                "operationId": "list_tasks_api_v1_tasks__get",
                "parameters": [
                    {
                        "name": "site_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Site Id"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Status"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TaskResponse"
                                    },
                                    "title": "Response List Tasks Api V1 Tasks  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "TaskResponse": {
                "properties": {
                    "task_id": {
                        "type": "string",
                        "title": "Task Id",
                        "description": "任务ID"
                    },
                    "site_id": {
                        "type": "string",
                        "title": "Site Id",
                        "description": "站点ID"
                    },
                    "status": {
                        "$ref": "#/components/schemas/TaskStatus",
                        "description": "任务状态"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At",
                        "description": "创建时间"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At",
                        "description": "更新时间"
                    },
                    "completed_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Completed At",
                        "description": "任务完成时间"
                    },
                    "msg": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Msg",
                        "description": "错误信息"
                    },
                    "error_details": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error Details",
                        "description": "详细错误信息"
                    },
                    "task_metadata": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Task Metadata",
                        "description": "任务元数据"
                    },
                    "system_info": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "System Info",
                        "description": "系统信息"
                    },
                    "duration": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Duration",
                        "description": "任务持续时间(秒)"
                    }
                },
                "type": "object",
                "required": [
                    "task_id",
                    "site_id",
                    "status",
                    "created_at",
                    "updated_at"
                ],
                "title": "TaskResponse",
                "description": "任务响应模型"
            },
            "TaskStatus": {
                "type": "string",
                "enum": [
                    "ready",
                    "pending",
                    "running",
                    "success",
                    "failed",
                    "cancelled"
                ],
                "title": "TaskStatus",
                "description": "任务状态枚举"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        }
    }
}